First pass at a timer based on motion
This was written for an Arduino 101 board.

Watch the accelerometer, and count up in seconds as long as no
movement occurs (given a tolerance, or "sensitivity").

When motion is detected, call a function to perform an action. (The
action will be to disable an alarm if one has sounded).

If the counter passes a set number of seconds, start making a
sound (not yet implemented).